requirements.txt
sqlalchemy
python-decouple
psycopg2-binary
pandas
postgres
fastapi
uvicorn
pydantic

